Poc: Add new "Satellite" Pollster and its Ganglia Implementation
This patch provide an Agent to deploy inside intances to allow Post
sample to Ceilometer API.

This new agent/pollster is build like others in Ceilometer and use/abuse
of the pipeline concept with its (new) HTTP publisher.
Note: obviously samples POSTed are not signed, the 'standrad' Keystone
authentification is used to Post samples.

Notice: an *ugly* patch amends the pipeline behavior by hijacking some
meanings .. need to redefine and find an common abstract to match with
master.

Part-of-Blueprint : applications-metrics-from-instances

Change-Id: I06ce7372f2b9f9ede48e94ed5115fd446e4aa040